# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Mural Maz Lince API",
    "description": "API REST del sistema Mural Maz Lince. Digitaliza el tablero de anuncios estudiantil de la Universidad Autónoma de Occidente, Unidad Regional Mazatlán.\n\n**Estado:** Definitivo Sprint 10 — incluye pipeline de moderación (Vision + bad-words + dHash), panel admin y endpoints operativos de shadowban/cola.\n\n**Autenticación:** Cookies HttpOnly con JWT (ADR-04). Incluir `credentials: 'include'` en cada solicitud del cliente.\n\n**Referencia SRS:** v1.4 | **Referencia SAD:** v1.3 | **Referencia DDC:** v1.2",
    "version": "1.0.0-sprint10",
    "contact": {
      "name": "Miriam Alanis — Directora del Proyecto",
      "url": "https://github.com/devHamialco/MuralMazLince"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Entorno local (docker-compose)"
    },
    {
      "url": "https://muralmaz.railway.app",
      "description": "Producción (Railway.app) — disponible al cierre del Sprint 14"
    }
  ],
...
  "paths": {
...
    "/announcements": {
      "get": {
        "tags": ["feed"],
        "summary": "Feed de anuncios activos (paginación por cursor)",
        "description": "Devuelve los 20 anuncios activos más recientes a partir del cursor indicado. Accesible sin autenticación (ROL-01). El campo `whatsapp_number` nunca aparece en la respuesta. **RF-08, RF-09, RNF-01, RNF-02, RN-09.**",
        "operationId": "getAnnouncements",
        "security": [],
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "description": "ID del último anuncio recibido. Omitir para la primera página.",
            "required": false,
            "schema": { "type": "integer", "example": 45 }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Número de anuncios por bloque. Valor fijo del sistema: 20.",
            "required": false,
            "schema": { "type": "integer", "default": 20, "maximum": 20 }
          },
          {
            "name": "category",
            "in": "query",
            "description": "Código de categoría para filtrar (ej: CAT-01). Solo disponible para ROL-02 o superior.",
            "required": false,
            "schema": { "type": "string", "example": "CAT-01" }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de anuncios paginada.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AnnouncementFeedResponse" }
              }
            }
          },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      },
      "post": {
        "tags": ["announcements"],
        "summary": "Crear nuevo anuncio",
        "description": "Crea un anuncio dentro de un proyecto del emprendedor autenticado. Ejecuta el pipeline de moderación completo en paralelo (Vision API + bad-words + dHash). Devuelve HTTP 201 si el anuncio queda activo o HTTP 202 si queda en revisión. **RF-19, RF-22, RF-26, RF-27, RF-28, RF-31, RF-32, RN-04, RN-05, RN-11, RN-12.**",
        "operationId": "createAnnouncement",
        "security": [{ "cookieAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": { "$ref": "#/components/schemas/CreateAnnouncementRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Anuncio publicado y activo en el feed.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AnnouncementDetail" }
              }
            }
          },
          "202": {
            "description": "Anuncio creado pero en revisión pendiente. La moderación automática generó una alerta.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AnnouncementDetail" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "413": { "$ref": "#/components/responses/PayloadTooLarge" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
    "/announcements/{id}": {
      "get": {
        "tags": ["feed"],
        "summary": "Detalle de un anuncio",
        "description": "Devuelve la información completa de un anuncio. Si el usuario está autenticado, incluye el campo `contact_link` con la URL wa.me generada dinámicamente. Para ROL-01, `contact_link` es `null`. El número de WhatsApp nunca se expone. **RF-10, RF-11, RNF-04, RN-09.**",
        "operationId": "getAnnouncementById",
        "security": [],
        "parameters": [
          { "$ref": "#/components/parameters/AnnouncementId" }
        ],
        "responses": {
          "200": {
            "description": "Detalle del anuncio.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AnnouncementDetail" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      },
      "patch": {
        "tags": ["announcements"],
        "summary": "Editar anuncio",
        "description": "Edita título, descripción, categoría o vigencia de un anuncio activo del emprendedor autenticado. Si se incluye una nueva imagen, el pipeline de moderación se re-ejecuta completo. **RF-20, RN-05.**",
        "operationId": "updateAnnouncement",
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/AnnouncementId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": { "$ref": "#/components/schemas/UpdateAnnouncementRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Anuncio actualizado.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AnnouncementDetail" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "413": { "$ref": "#/components/responses/PayloadTooLarge" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      },
      "delete": {
        "tags": ["announcements"],
        "summary": "Eliminar anuncio",
        "description": "Elimina el anuncio y su imagen de Cloudinary usando `cloudinary_id` (no la URL). Requiere confirmación explícita del cliente. **RF-21, RF-33, RN-05.**",
        "operationId": "deleteAnnouncement",
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/AnnouncementId" }
        ],
        "responses": {
          "200": {
            "description": "Anuncio e imagen eliminados correctamente.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MessageResponse" },
                "example": { "message": "Anuncio eliminado correctamente." }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token",
        "description": "JWT en cookie HttpOnly. Emitido por POST /auth/login. No accesible desde JavaScript del cliente (ADR-04). Incluir `credentials: 'include'` en cada solicitud fetch del frontend."
      }
    },
...
    "parameters": {
      "AnnouncementId": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "ID del anuncio.",
        "schema": { "type": "integer", "example": 42 }
      },
...
    },
    "schemas": {
...
      "MessageResponse": {
        "type": "object",
        "required": ["message"],
        "properties": {
          "message": { "type": "string" }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": { "type": "string", "description": "Descripción del error. Nunca expone datos sensibles." }
        }
      },
...
      "Category": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "code": { "type": "string", "example": "CAT-01" },
          "name": { "type": "string", "example": "Alimentos y bebidas" },
          "group_name": { "type": "string", "example": "Productos físicos" },
          "is_active": { "type": "boolean", "example": true }
        }
      },
      "AnnouncementSummary": {
        "type": "object",
        "description": "Vista resumida del anuncio para el feed. RF-08.",
        "properties": {
          "id": { "type": "integer", "example": 42 },
          "title": { "type": "string", "example": "Diseño de logos profesionales" },
          "cloudinary_url": { "type": "string", "format": "uri", "example": "https://res.cloudinary.com/mural/image/upload/v1/mural_lince/abc123.jpg" },
          "category": { "$ref": "#/components/schemas/Category" },
          "project_name": { "type": "string", "example": "Tech Lince — Diseño Gráfico" },
          "entrepreneur_display_name": {
            "type": "string",
            "description": "Proveniente de entrepreneur_profiles.display_name. RF-08, RF-39.",
            "example": "Mariana G. Tech Lince"
          },
          "total_likes": { "type": "integer", "example": 14 },
          "average_rating": { "type": "number", "format": "float", "example": 2.7 },
          "total_ratings": { "type": "integer", "example": 11 },
          "expires_at": { "type": "string", "format": "date-time", "example": "2026-05-15T23:59:59Z" },
          "created_at": { "type": "string", "format": "date-time", "example": "2026-04-10T14:30:00Z" }
        }
      },
      "AnnouncementDetail": {
        "type": "object",
        "description": "Vista completa del anuncio. Incluye contact_link solo para usuarios autenticados. RF-10, RF-11.",
        "allOf": [
          { "$ref": "#/components/schemas/AnnouncementSummary" },
          {
            "type": "object",
            "properties": {
              "description": {
                "type": "string",
                "maxLength": 500,
                "example": "Diseño de logotipos, tarjetas de presentación y material visual para tu emprendimiento."
              },
              "status": {
                "type": "string",
                "enum": ["active", "pending_review", "rejected", "shadowban"],
                "example": "active"
              },
              "contact_link": {
                "type": "string",
                "format": "uri",
                "nullable": true,
                "description": "URL wa.me generada dinámicamente en el servidor. null para ROL-01. El número nunca se expone. RF-11, RN-09.",
                "example": "https://wa.me/5216691234567"
              },
              "project_id": { "type": "integer", "example": 7 },
              "custom_category": {
                "type": "string",
                "nullable": true,
                "description": "Solo presente si category.code = CAT-00 (Otro). RN-12.",
                "example": null
              }
            }
          }
        ]
      },
      "AnnouncementFeedResponse": {
        "type": "object",
        "description": "Respuesta paginada del feed por cursor. RNF-02.",
        "properties": {
          "data": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/AnnouncementSummary" }
          },
          "next_cursor": {
            "type": "integer",
            "nullable": true,
            "description": "ID del último anuncio devuelto. Enviar como parámetro `cursor` en la siguiente solicitud. null si no hay más anuncios.",
            "example": 22
          },
          "has_more": { "type": "boolean", "example": true }
        }
      },
      "CreateAnnouncementRequest": {
        "type": "object",
        "description": "Multipart/form-data. La imagen se comprime en el cliente antes del envío (RF-31).",
        "required": ["project_id", "title", "image", "category_id", "expires_at"],
        "properties": {
          "project_id": { "type": "integer", "example": 7 },
          "title": { "type": "string", "maxLength": 120, "example": "Diseño de logos profesionales" },
          "description": { "type": "string", "maxLength": 500, "example": "Logos únicos para tu marca." },
          "image": {
            "type": "string",
            "format": "binary",
            "description": "Imagen comprimida. Máx. 500 KB / 1,200 px (RF-31, RNF-03)."
          },
          "category_id": { "type": "integer", "description": "ID de la categoría predefinida.", "example": 9 },
          "custom_category": {
            "type": "string",
            "maxLength": 80,
            "nullable": true,
            "description": "Requerido solo si category_id corresponde a CAT-00 (Otro). Sujeto a filtro bad-words. RN-12.",
            "example": null
          },
          "expires_at": {
            "type": "string",
            "format": "date",
            "description": "Fecha de vencimiento. Mínimo mañana, máximo 365 días desde hoy. RF-22.",
            "example": "2026-05-15"
          }
        }
      },
      "UpdateAnnouncementRequest": {
        "type": "object",
        "description": "Todos los campos son opcionales. Si se incluye `image`, el pipeline de moderación se re-ejecuta. RF-20.",
        "properties": {
          "title": { "type": "string", "maxLength": 120 },
          "description": { "type": "string", "maxLength": 500 },
          "image": { "type": "string", "format": "binary" },
          "category_id": { "type": "integer" },
          "custom_category": { "type": "string", "maxLength": 80, "nullable": true },
          "expires_at": { "type": "string", "format": "date" }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "400 Bad Request — Checksum inválido, campos faltantes o formato de WhatsApp incorrecto.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "error": "El formato de la matrícula es inválido. Debe tener 8 dígitos y comenzar con 2." }
          }
        }
      },
      "Unauthorized": {
        "description": "401 Unauthorized — Cookie de sesión ausente o expirada.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "error": "Sesión no encontrada. Por favor inicia sesión." }
          }
        }
      },
      "Forbidden": {
        "description": "403 Forbidden — Cuenta suspendida, o rol insuficiente para la acción solicitada.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "error": "Tu cuenta está suspendida. Contacta al administrador." }
          }
        }
      },
      "NotFound": {
        "description": "404 Not Found — El recurso solicitado no existe.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "error": "Anuncio no encontrado." }
          }
        }
      },
      "Conflict": {
        "description": "409 Conflict — Matrícula ya registrada, reporte duplicado o límite de proyectos/anuncios alcanzado.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "error": "Esta matrícula ya está registrada en el sistema." }
          }
        }
      },
      "PayloadTooLarge": {
        "description": "413 Payload Too Large — La imagen supera 500 KB o 1,200 px. Segunda línea de defensa server-side (RF-31, RNF-03).",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "error": "La imagen supera el tamaño máximo permitido (500 KB / 1,200 px)." }
          }
        }
      },
...
      "InternalServerError": {
        "description": "500 Internal Server Error — Error inesperado en el servidor. El detalle se registra en el log; el cliente recibe un mensaje genérico.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "error": "Ocurrió un error interno. Por favor intenta más tarde." }
          }
        }
      }
    }
  }
}